Dequeue

Description

This is a module that implements a dequeue and it's utilities.

Documentation

data Dequeue a

Type that represents a dequeue.

Constructors

Dequeue [a]

emptyDEQ :: Dequeue a

Function that returns an empty dequeue.

isEmptyDEQ :: Dequeue a -> Bool

Function that takes a dequeue and returns True if it is empty or Falese if it is not empty.

lengthDEQ :: Dequeue a -> Int

Function that takes a dequeue and returns it length.

firstDEQ :: Dequeue a -> Maybe a

Function that takes a dequeue and returns the first element of it.

lastDEQ :: Dequeue a -> Maybe a

Function that takes a dequeue and returns the last element of it.

takeFrontDEQ :: Int -> Dequeue a -> [a]

Function that takes a number n and returns an array of first n elements in order they would be poped. If dequeue's length is smaller then n, it returns array made out of all elements of the dequeue.

takeBackDEQ :: Int -> Dequeue a -> [a]

Function that takes a number n and returns an array of last n elements in order they would be poped. If dequeue's length is smaller then n, it returns array made out of all elements of the dequeue.

pushFrontDEQ :: Dequeue a -> a -> Dequeue a

Function that takes a dequeue and an element and adds it at the beginning of the dequeue.

popFrontDEQ :: Dequeue a -> Maybe (a, Dequeue a)

Function that takes a dequeue and returns a Just pair of the first element and the dequeue without the popped element, if dequeue is empty it returns Nothing.

pushBackDEQ :: Dequeue a -> a -> Dequeue a

Function that takes a dequeue and an element and adds it at the end of the dequeue.

popBackDEQ :: Dequeue a -> Maybe (a, Dequeue a)

Function that takes a dequeue and returns a Just pair of the last element and the dequeue without the popped element, if dequeue is empty it returns Nothing.

fromListDEQ :: [a] -> Dequeue a

Function that takes a list and returns a dequeue made out of that list.