Heap

Description

This is a modulte that is implementing a pairing heap and it's utilities.

Documentation

data Heap a

Type that represents a Heap, it can be Empty, or it can have node that has a list of it's chirldren.

Constructors

EmptyH

Node a [Heap a]

clearH :: Heap a -> Heap a

Fuction that takes a heap and return empty heap.

emptyHe :: Heap a

Function that returns an empty heap.

isEmptyHH :: Heap a -> Bool

Function that takes a dequeue and returns True if it is empty or Falese if it is not empty.

pushH :: Ord a => a -> Heap a -> Heap a

Function that takes an element and a heap and returns a heap with this element added.

popH :: Ord a => Heap a -> Maybe (a, Heap a)

Function that takes a heap and returns a Just pair of the first element and the heap without the element, if heap is empty it returns Nothing.

topH :: Heap a -> Maybe a

Function that takes a reap and returns Just the element that is on the top of the heap (the smallest element).

heapSort :: Ord a => [a] -> [a]

Function that takes an array and returns a sotred array made out of elements of taken array. It uses heap to sort the array.

makeHeapFromArray :: Ord a => [a] -> Heap a

Function that takes an array and returns heap made out of the elements of taken array.

makeArrayFromHeap :: Ord a => Heap a -> [a]

Function that takes a heap and returns a array made out of the heap. It is sorted increasingly.