Lib

Description

a tu troche wiecej powinno byc napisane o tym module, ale nic wiecej nie napisze, tylko pokaze to some markup

Some functions

someFunc

Arguments

:: IO ()
wejscie/wyjscie

wyswietla nam wynik naszego szyfrowania

cSzyfruj2

Arguments

:: Functor f

=> f Char
ciag wejsciowy

-> Int
liczba szyfrujaca

-> f Char
ciag znakow