Message

Documentation

data Message

Define a type Message, each has a message of type String, a senderid of type Int and a receiverid of type Int

Constructors

Message

Fields

message :: String

senderId :: Int

receiverId :: Int

sendMessage :: Int -> MVar [Message] -> IO ()

Start sending message

countMessage :: [Message] -> User -> IO ()

Count the total number of messages each user received