Queue

Description

This is a module that implements a queue and it's utilities.

Documentation

data Queue a

Type that represents a queue.

Constructors

Queue [a]

emptyQ :: Queue a

Function that returns an empty queue.

isEmptyQ :: Queue a -> Bool

Function tht takes a queue and returns True if it is emty or False if it is not empty.

addQ :: a -> Queue a -> Queue a

Function that takes an element and adds it to the queue.

remQ :: Queue a -> Maybe (a, Queue a)

Function that takes a queue a returns Just pair of the first element and the queue without this element, If queue is empty it returns Nothing.

takeFrontQ :: Int -> Queue a -> [a]

Function that takes a number n and returns an array of first n elements in order they would be removed. If queue's length is smaller then n, it returns array made out of all elements of the queue.

fromListQ :: [a] -> Queue a

Function that takes a list and returns a queue made out of that list.