User

Documentation

data User

Define a type User, each user has a username of type String and a userid of type Int

Constructors

User

Fields

userName :: String

userId :: Int

userList :: [User]

Generate a list of 10 users